+++ docs/changelog.md
@@ -6,6 +6,7 @@ Unreleased
 - Make StructureMap/$transform always available, not only on onlyOneEngine mode (#428)
 - Fix StructureMap/$transform not reading the `source` parameter from the Parameters body
 - Fix the rendering of the StructureMap/$transform operation in the GUI
+- Don't copy the suppressedWarnInfos/suppressedErrors from the main engine to the validation engines (#539)
 
 2026/08/14 Release 4.1.13